Support locking through lockfile or django-locking
To prevent concurrent sending of messages a lock is acquired. To allow for multiple workers on different machines, you can now opt to use a database lock instead of a file lock. Which kind of lock is used depends on the `SMSGATEWAY_LOCKING_CLASS` setting, with a default value of `smsgateway.locking.files.Lock`